@@ -42,67 +42,69 @@
     "build:sassdoc:en:staging": "set NODE_ENV=staging && set SASSDOC_LANG=en && gulp sassdocBuildEN",
     "build:sassdoc:ja:production": "set NODE_ENV=production && set SASSDOC_LANG=jp && gulp sassdocBuildJA --render true",
     "build:sassdoc:ja:staging": "set NODE_ENV=staging && set SASSDOC_LANG=jp && gulp sassdocBuildJA --render true",
+    "build:i18n": "tsc --project projects/igniteui-angular-i18n/tsconfig.build.json",
+    "lint:i18n": "eslint projects/igniteui-angular-i18n/src/**/*.ts",
+    "lint:i18n:dist": "eslint -c projects/igniteui-angular-i18n/.eslintrc.dist.i18n.json dist/igniteui-angular-i18n/**/*.d.ts --no-eslintrc",
     "postinstall": "gulp copyGitHooks"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.0.0",
-    "@angular/common": "^12.0.0",
-    "@angular/compiler": "^12.0.0",
-    "@angular/core": "^12.0.0",
-    "@angular/forms": "^12.0.0",
-    "@angular/platform-browser": "^12.0.0",
-    "@angular/platform-browser-dynamic": "^12.0.0",
-    "@angular/router": "^12.0.0",
-    "@igniteui/material-icons-extended": "^2.7.0",
+    "@angular/animations": "^12.1.1",
+    "@angular/common": "^12.1.1",
+    "@angular/compiler": "^12.1.1",
+    "@angular/core": "^12.1.1",
+    "@angular/forms": "^12.1.1",
+    "@angular/platform-browser": "^12.1.1",
+    "@angular/platform-browser-dynamic": "^12.1.1",
+    "@angular/router": "^12.1.1",
+    "@igniteui/material-icons-extended": "^2.8.0",
     "@juggle/resize-observer": "^3.3.1",
-    "@types/hammerjs": "^2.0.39",
+    "@types/hammerjs": "^2.0.40",
     "@types/source-map": "0.5.2",
     "classlist.js": "^1.1.20150312",
     "core-js": "^2.6.11",
     "hammerjs": "^2.0.8",
-    "igniteui-angular-i18n": "^12.0.3",
     "igniteui-trial-watermark": "^1.0.3",
     "jszip": "^3.6.0",
     "karma-parallel": "^0.3.1",
     "lodash.mergewith": "^4.6.2",
     "rxjs": "^6.6.7",
     "setimmediate": "^1.0.5",
-    "tslib": "^2.2.0",
+    "tslib": "^2.3.0",
     "uuid": "^8.3.0",
     "web-animations-js": "^2.3.2",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.0.0",
-    "@angular-devkit/schematics": "^12.0.0",
-    "@angular-eslint/builder": "^4.0.0",
-    "@angular-eslint/eslint-plugin": "^4.0.0",
-    "@angular-eslint/eslint-plugin-template": "^4.0.0",
-    "@angular-eslint/schematics": "^12.0.0",
-    "@angular-eslint/template-parser": "^4.0.0",
-    "@angular/cli": "~12.0.0",
-    "@angular/compiler-cli": "^12.0.0",
-    "@angular/language-service": "^12.0.0",
+    "@angular-devkit/build-angular": "~12.1.1",
+    "@angular-devkit/schematics": "^12.1.1",
+    "@angular-eslint/builder": "^12.2.0",
+    "@angular-eslint/eslint-plugin": "^12.2.0",
+    "@angular-eslint/eslint-plugin-template": "^12.2.0",
+    "@angular-eslint/schematics": "^12.2.0",
+    "@angular-eslint/template-parser": "^12.2.0",
+    "@angular/cli": "~12.1.1",
+    "@angular/compiler-cli": "^12.1.1",
+    "@angular/language-service": "^12.1.1",
     "@angularclass/hmr": "^2.1.3",
-    "@types/jasmine": "^3.7.2",
-    "@types/jasminewd2": "^2.0.9",
-    "@types/node": "^12.20.13",
-    "@types/webpack-env": "^1.15.2",
+    "@types/jasmine": "^3.7.8",
+    "@types/jasminewd2": "^2.0.10",
+    "@types/node": "^12.20.15",
+    "@types/webpack-env": "^1.16.2",
     "@typescript-eslint/eslint-plugin": "4.16.1",
     "@typescript-eslint/parser": "4.16.1",
-    "browser-sync": "^2.26.12",
+    "browser-sync": "^2.27.4",
     "codelyzer": "^6.0.2",
-    "coveralls": "^3.1.0",
-    "eslint": "^7.26.0",
+    "coveralls": "^3.1.1",
+    "eslint": "^7.30.0",
     "eslint-plugin-import": "2.22.1",
     "eslint-plugin-jsdoc": "30.7.6",
     "eslint-plugin-prefer-arrow": "1.2.2",
     "gulp": "^4.0.2",
     "gulp-cached": "^1.1.1",
     "gulp-concat": "^2.6.1",
-    "gulp-dart-sass": "^1.0.2",
     "gulp-postcss": "^9.0.0",
+    "gulp-sass": "^5.0.0",
     "gulp-shell": "^0.6.5",
     "gulp-sourcemaps": "^2.6.4",
     "gulp-typescript": "^5.0.0-alpha.3",
@@ -134,9 +137,9 @@
     "themeleon": "^3.0.2",
     "ts-node": "~7.0.1",
     "tslint": "~6.1.0",
-    "typedoc": "^0.20.36",
+    "typedoc": "^0.20.37",
     "typedoc-plugin-localization": "^2.2.4",
-    "typescript": "^4.2.4",
+    "typescript": "^4.3.5",
     "webpack-sources": "1.3.0"
   }
 }
