@@ -4,7 +4,7 @@ const autoprefixer = require('autoprefixer');
 const browserSync = require('browser-sync').create();
 const del = require('del');
 const gulp = require('gulp');
-const sass = require('gulp-dart-sass');
+const sass = require('gulp-sass')(require('sass'));
 const sourcemaps = require('gulp-sourcemaps');
 const postcss = require('gulp-postcss');
 const process = require('process');
@@ -17,10 +17,8 @@ const { series } = require('gulp');
 const {spawnSync} = require('child_process');
 const slash = require('slash');
 
-sass.compiler = require('sass');
-
 const STYLES = {
-    SRC: './projects/igniteui-angular/src/lib/core/styles/themes/presets/*',
+    SRC: './projects/igniteui-angular/src/lib/core/styles/themes/presets/**/*',
     DIST: './dist/igniteui-angular/styles',
     MAPS: './maps',
     THEMING: {
@@ -39,7 +37,7 @@ const TYPEDOC_THEME = {
     OUTPUT: slash(path.join(DOCS_OUTPUT_PATH, 'typescript'))
 };
 
-module.exports.buildStyle =  (cb) => {
+module.exports.buildStyle = () => {
     const prefixer = postcss([autoprefixer({
         cascade: false,
         grid: true
@@ -50,7 +48,7 @@ module.exports.buildStyle =  (cb) => {
 
     const myEventEmitter = new EventEmitter();
 
-    gulp.src(STYLES.SRC)
+    return gulp.src(STYLES.SRC)
         .pipe(sourcemaps.init())
         .pipe(sass.sync(STYLES.CONFIG).on('error', err => {
             sass.logError.bind(myEventEmitter)(err);
@@ -59,9 +57,7 @@ module.exports.buildStyle =  (cb) => {
         }))
         .pipe(prefixer)
         .pipe(sourcemaps.write(STYLES.MAPS))
-        .pipe(gulp.dest(STYLES.DIST));
-
-    cb();
+        .pipe(gulp.dest(STYLES.DIST))
 };
 
 module.exports.copyGitHooks = async (cb) => {
