@@ -66,8 +66,7 @@ export class IgxCheckboxComponent implements ControlValueAccessor, EditorProvide
      * Provides references to the `IgxCheckboxComponent` and the `checked` property as event arguments.
      */
     // eslint-disable-next-line @angular-eslint/no-output-native
-    @Output()
-    public readonly change: EventEmitter<IChangeCheckboxEventArgs> = new EventEmitter<IChangeCheckboxEventArgs>();
+    @Output() public readonly change: EventEmitter<IChangeCheckboxEventArgs> = new EventEmitter<IChangeCheckboxEventArgs>();
     /**
      * Returns reference to the native checkbox element.
      *
