@@ -253,8 +253,8 @@ export class IgxFilteringService implements OnDestroy {
         const grid = this.grid;
         const filteringTree = grid.filteringExpressionsTree;
         grid.crudService.endEdit(false);
-        if (grid.paging) {
-            grid.page = 0;
+        if (grid.paginator) {
+            grid.paginator.page = 0;
         }
 
         filteringTree.filteringOperands = [];
@@ -346,8 +346,8 @@ export class IgxFilteringService implements OnDestroy {
         }
 
         this.grid.crudService.endEdit(false);
-        if (grid.paging) {
-            grid.page = 0;
+        if (grid.paginator) {
+            grid.paginator.page = 0;
         }
         grid.filteringExpressionsTree = newFilteringTree;
 
@@ -560,8 +560,8 @@ export class IgxFilteringService implements OnDestroy {
         const filteringTree = grid.filteringExpressionsTree;
         this.grid.crudService.endEdit(false);
 
-        if (grid.paging) {
-            grid.page = 0;
+        if (grid.paginator) {
+            grid.paginator.page = 0;
         }
 
         const fieldFilterIndex = filteringTree.findIndex(fieldName);
