@@ -226,11 +226,14 @@ export class IgxGridStateDirective {
                 return { paging: pagingState };
             },
             restoreFeatureState: (context: IgxGridStateDirective, state: IPagingState): void => {
-                if (context.currGrid.perPage !== state.recordsPerPage) {
-                    context.currGrid.perPage = state.recordsPerPage;
+                if (!context.currGrid.paginator) {
+                    return;
+                }
+                if (context.currGrid.paginator.perPage !== state.recordsPerPage) {
+                    context.currGrid.paginator.perPage = state.recordsPerPage;
                     context.currGrid.cdr.detectChanges();
                 }
-                context.currGrid.page = state.index;
+                context.currGrid.paginator.page = state.index;
             }
         },
         rowSelection: {
