@@ -338,7 +338,7 @@ describe('IgxTreeGrid - Expanding / Collapsing #tGrid', () => {
 
         it('should update current page when \'collapseAll\' ', fakeAsync(() => {
             // Test prerequisites
-            treeGrid.paging = true;
+            fix.detectChanges();
             treeGrid.perPage = 4;
             fix.detectChanges();
             tick(16);
@@ -458,7 +458,6 @@ describe('IgxTreeGrid - Expanding / Collapsing #tGrid', () => {
 
         it('Should update the paginator when navigating through pages', () => {
             // Test prerequisites
-            treeGrid.paging = true;
             fix.detectChanges();
 
             // Verify current page
@@ -787,7 +786,8 @@ describe('IgxTreeGrid - Expanding / Collapsing #tGrid', () => {
 
         it('should update current page when \'collapseAll\' ', fakeAsync(() => {
             // Test prerequisites
-            treeGrid.paging = true;
+            fix.componentInstance.paging = true;
+            fix.detectChanges();
             treeGrid.perPage = 2;
             fix.detectChanges();
             tick(16);
@@ -818,7 +818,8 @@ describe('IgxTreeGrid - Expanding / Collapsing #tGrid', () => {
 
         it('Should update the paginator when a row of any level is expanded', fakeAsync(() => {
             // Test prerequisites
-            treeGrid.paging = true;
+            fix.componentInstance.paging = true;
+            fix.detectChanges();
             treeGrid.perPage = 5;
             fix.detectChanges();
             tick(16);
@@ -863,7 +864,8 @@ describe('IgxTreeGrid - Expanding / Collapsing #tGrid', () => {
 
         it('Should update the paginator when a row of any level is collapsed', fakeAsync(() => {
             // Test prerequisites
-            treeGrid.paging = true;
+            fix.componentInstance.paging = true;
+            fix.detectChanges();
             treeGrid.perPage = 5;
             fix.detectChanges();
             tick(16);
@@ -893,7 +895,8 @@ describe('IgxTreeGrid - Expanding / Collapsing #tGrid', () => {
 
         it('Should update the paginator when navigating through pages', fakeAsync(() => {
             // Test prerequisites
-            treeGrid.paging = true;
+            fix.componentInstance.paging = true;
+            fix.detectChanges();
             treeGrid.perPage = 5;
             fix.detectChanges();
             tick(16);
