@@ -1625,7 +1625,8 @@ describe('IgxTreeGrid - Integration #tGrid', () => {
             treeGrid.pinRow(147);
             fix.detectChanges();
 
-            treeGrid.paging = true;
+            fix.componentInstance.paging = true;
+            fix.detectChanges();
             treeGrid.perPage = 5;
             fix.detectChanges();
 
@@ -1638,7 +1639,8 @@ describe('IgxTreeGrid - Integration #tGrid', () => {
         });
 
         it('should correctly apply paging state for grid and paginator when there are pinned rows.', fakeAsync(() => {
-            treeGrid.paging = true;
+            fix.componentInstance.paging = true;
+            fix.detectChanges();
             treeGrid.perPage = 3;
             treeGrid.height = '700px';
             fix.detectChanges();
@@ -1665,7 +1667,8 @@ describe('IgxTreeGrid - Integration #tGrid', () => {
         }));
 
         it('should have the correct records shown for pages with pinned rows', () => {
-            treeGrid.paging = true;
+            fix.componentInstance.paging = true;
+            fix.detectChanges();
             treeGrid.perPage = 6;
             treeGrid.height = '700px';
             fix.detectChanges();
