@@ -16,6 +16,8 @@ import { IgxTreeGridRowComponent } from './tree-grid-row.component';
 import { GridSummaryFunctions, GridFunctions } from '../../test-utils/grid-functions.spec';
 import { IgxNumberFilteringOperand } from '../../data-operations/filtering-condition';
 import { DebugElement } from '@angular/core';
+import { IgxTreeGridComponent } from './tree-grid.component';
+import { IgxSummaryRow, IgxTreeGridRow } from '../grid-public-row';
 
 describe('IgxTreeGrid - Summaries #tGrid', () => {
     configureTestSuite();
@@ -39,7 +41,7 @@ describe('IgxTreeGrid - Summaries #tGrid', () => {
 
     describe('', () => {
         let fix;
-        let treeGrid;
+        let treeGrid: IgxTreeGridComponent;
         beforeEach(waitForAsync(/** height/width setter rAF */() => {
             fix = TestBed.createComponent(IgxTreeGridSummariesKeyComponent);
             fix.detectChanges();
@@ -53,6 +55,23 @@ describe('IgxTreeGrid - Summaries #tGrid', () => {
             // Expand second row and verify summaries
             treeGrid.toggleRow(treeGrid.getRowByIndex(1).key);
             fix.detectChanges();
+
+            const secondRow = treeGrid.getRowByIndex(1);
+            const thirdRow = treeGrid.getRowByIndex(2);
+            const summaryRow = treeGrid.getRowByIndex(4);
+
+            // First row is IgxTreeRow 4thRow is IgxSummaryRow
+            expect(secondRow instanceof IgxTreeGridRow).toBe(true);
+            expect(thirdRow instanceof IgxTreeGridRow).toBe(true);
+            expect(secondRow.index).toBe(1);
+            expect(secondRow.viewIndex).toBe(1);
+            expect(thirdRow.index).toBe(2);
+            expect(thirdRow.viewIndex).toBe(2);
+
+            expect(thirdRow.parent.data).toBe(secondRow.data);
+            expect(secondRow.children[0].data).toBe(thirdRow.data);
+
+            expect(summaryRow instanceof IgxSummaryRow).toBe(true);
 
             verifyTreeBaseSummaries(fix);
             verifySummaryForRow847(fix, 4);
@@ -82,6 +101,16 @@ describe('IgxTreeGrid - Summaries #tGrid', () => {
             verifyTreeBaseSummaries(fix);
             verifySummaryForRow147(fix, 1);
             expect(GridSummaryFunctions.getAllVisibleSummariesLength(fix)).toEqual(2);
+
+            const firstRow = treeGrid.getRowByIndex(0);
+            const summaryRow = treeGrid.getRowByIndex(1);
+
+            // First row is IgxTreeRow 4thRow is IgxSummaryRow
+            expect(firstRow instanceof IgxTreeGridRow).toBe(true);
+            expect(firstRow.index).toBe(0);
+            expect(firstRow.viewIndex).toBe(0);
+
+            expect(summaryRow instanceof IgxSummaryRow).toBe(true);
 
             // Expand second row and verify summaries
             treeGrid.toggleRow(treeGrid.getRowByIndex(5).key);
@@ -172,11 +201,21 @@ describe('IgxTreeGrid - Summaries #tGrid', () => {
             treeGrid.showSummaryOnCollapse = true;
             fix.detectChanges();
 
+            let secondRow = treeGrid.getRowByIndex(1);
+            expect(secondRow.index).toEqual(1);
+            expect(secondRow.viewIndex).toEqual(1);
+            expect(secondRow instanceof IgxSummaryRow).toBe(true);
+
             summaries = GridSummaryFunctions.getAllVisibleSummaries(fix);
             expect(summaries.length).toBe(4);
 
             treeGrid.showSummaryOnCollapse = false;
             fix.detectChanges();
+
+            secondRow = treeGrid.getRowByIndex(1);
+            expect(secondRow.index).toEqual(1);
+            expect(secondRow.viewIndex).toEqual(1);
+            expect(secondRow instanceof IgxSummaryRow).toBe(false);
 
             summaries = GridSummaryFunctions.getAllVisibleSummaries(fix);
             expect(summaries.length).toBe(0);
@@ -199,6 +238,11 @@ describe('IgxTreeGrid - Summaries #tGrid', () => {
 
             treeGrid.toggleRow(treeGrid.getRowByIndex(3).key);
             fix.detectChanges();
+
+						const gridSummaryRow = treeGrid.getRowByIndex(4);
+						expect(gridSummaryRow.index).toEqual(4);
+						expect(gridSummaryRow.viewIndex).toEqual(4);
+						expect(gridSummaryRow instanceof IgxSummaryRow).toBe(true);
 
             summaries = GridSummaryFunctions.getAllVisibleSummaries(fix);
             expect(summaries.length).toBe(4);
@@ -513,7 +557,8 @@ describe('IgxTreeGrid - Summaries #tGrid', () => {
         }));
 
         it('Paging: should render correct summaries when paging is enable and position is bottom', fakeAsync(() => {
-            treeGrid.paging = true;
+            fix.componentInstance.paging = true;
+            fix.detectChanges();
             treeGrid.perPage = 4;
             fix.detectChanges();
             tick(16);
@@ -524,6 +569,11 @@ describe('IgxTreeGrid - Summaries #tGrid', () => {
             treeGrid.toggleRow(treeGrid.getRowByIndex(0).key);
             fix.detectChanges();
             tick(16);
+
+            let summaryRow = treeGrid.getRowByIndex(4);
+            expect(summaryRow.index).toEqual(4);
+            expect(summaryRow.viewIndex).toEqual(4);
+            expect(summaryRow instanceof IgxSummaryRow).toBe(true);
 
             expect(GridSummaryFunctions.getAllVisibleSummariesLength(fix)).toEqual(2);
             verifyTreeBaseSummaries(fix);
@@ -539,6 +589,15 @@ describe('IgxTreeGrid - Summaries #tGrid', () => {
             fix.detectChanges();
             tick(16);
 
+            // TODO FIX
+            const firstRow = treeGrid.getRowByIndex(0);
+            summaryRow = treeGrid.getRowByIndex(2);
+            expect(firstRow.index).toEqual(0);
+            expect(firstRow.viewIndex).toEqual(4);
+            expect(summaryRow.index).toEqual(2);
+            expect(summaryRow.viewIndex).toEqual(6);
+            expect(summaryRow instanceof IgxSummaryRow).toBe(true);
+
             expect(GridSummaryFunctions.getAllVisibleSummariesLength(fix)).toEqual(3);
             verifyTreeBaseSummaries(fix);
             verifySummaryForRow147(fix, 3);
@@ -546,7 +605,8 @@ describe('IgxTreeGrid - Summaries #tGrid', () => {
         }));
 
         it('Paging: should render correct summaries when paging is enable and position is top', fakeAsync(() => {
-            treeGrid.paging = true;
+            fix.componentInstance.paging = true;
+            fix.detectChanges();
             treeGrid.perPage = 4;
             treeGrid.summaryPosition = 'top';
             fix.detectChanges();
@@ -573,11 +633,20 @@ describe('IgxTreeGrid - Summaries #tGrid', () => {
             fix.detectChanges();
             tick(16);
 
+            const firstRow = treeGrid.getRowByIndex(0);
+            expect(firstRow.index).toEqual(0);
+            expect(firstRow.viewIndex).toEqual(5);
+
             expect(GridSummaryFunctions.getAllVisibleSummariesLength(fix)).toEqual(1);
             verifyTreeBaseSummaries(fix);
 
             treeGrid.toggleRow(treeGrid.getRowByIndex(2).key);
             fix.detectChanges();
+
+            const summaryRow = treeGrid.getRowByIndex(3);
+            expect(summaryRow.index).toEqual(3);
+            expect(summaryRow.viewIndex).toEqual(8);
+            expect(summaryRow instanceof IgxSummaryRow).toBe(true);
 
             expect(GridSummaryFunctions.getAllVisibleSummariesLength(fix)).toEqual(2);
             verifySummaryForRow847(fix, 3);
