@@ -37,7 +37,7 @@ import { IgxGridNavigationService } from '../grid-navigation.service';
 import { GridType } from '../common/grid.interface';
 import { IgxColumnComponent } from '../columns/column.component';
 import { IgxTreeGridSelectionService } from './tree-grid-selection.service';
-import { GridSelectionMode } from '../common/enums';
+import { GridInstanceType, GridSelectionMode } from '../common/enums';
 import { IgxSummaryRow, IgxTreeGridRow } from '../grid-public-row';
 import { RowType } from '../common/row.interface';
 import { IgxGridCRUDService } from '../common/crud.service';
@@ -749,13 +749,13 @@ export class IgxTreeGridComponent extends IgxGridBaseDirective implements GridTy
             record = this.processedRecords.get(rowID);
             this._gridAPI.expand_path_to_record(record);
 
-            if (this.paging) {
+            if (this.paginator) {
                 const rowIndex = this.processedExpandedFlatData.indexOf(rowData);
-                const page = Math.floor(rowIndex / this.perPage);
+                const page = Math.floor(rowIndex / this.paginator.perPage);
 
-                if (this.page !== page) {
+                if (this.paginator.page !== page) {
                     delayScrolling = true;
-                    this.page = page;
+                    this.paginator.page = page;
                 }
             }
         }
@@ -798,7 +798,7 @@ export class IgxTreeGridComponent extends IgxGridBaseDirective implements GridTy
         const rec: any = this.dataView[index];
 
         if (this.isSummaryRecord(rec)) {
-            row = new IgxSummaryRow(this, index, rec.summaries);
+            row = new IgxSummaryRow(this, index, rec.summaries, GridInstanceType.TreeGrid);
         }
 
         if (!row && rec) {
