@@ -3,7 +3,7 @@ import { IgxPaginatorComponent } from './paginator.component';
 
 @Component({
     selector: 'igx-page-size',
-    templateUrl: 'page_size_selector.component.html',
+    templateUrl: 'page-size-selector.component.html',
 })
 export class IgxPageSizeSelectorComponent {
     /**
