@@ -20,10 +20,11 @@ export class GridDeclarationComponent {
 
 @Component({
     template: `
-        <igx-grid #grid1 [data]="data" [paging]="true" [perPage]="3">
+        <igx-grid #grid1 [data]="data">
             <igx-column field="ID"></igx-column>
             <igx-column field="Name"></igx-column>
             <igx-column field="JobTitle"></igx-column>
+            <igx-paginator [perPage]="5"></igx-paginator>
         </igx-grid>
     `
 })
