@@ -183,13 +183,14 @@ export class SelectionComponent extends BasicGridComponent {
 @Component({
     template: GridTemplateStrings.declareGrid(
         ` [width]="width" [height]="height" [rowSelection]="'multiple'" [primaryKey]="'ProductID'" [selectedRows]="selectedRows"`,
-        '', ColumnDefinitions.productBasicNumberID)
+        '', ColumnDefinitions.productBasicNumberID, '', '<igx-paginator *ngIf="paging"></igx-paginator>')
 })
 export class RowSelectionComponent extends BasicGridComponent {
     public data = SampleTestData.foodProductDataExtended();
     public width = '800px';
     public height = '600px';
     public selectedRows = [];
+    public paging = false;
 }
 
 @Component({
@@ -700,19 +701,21 @@ export class GridFeaturesComponent extends BasicGridComponent {
 @Component({
     template: GridTemplateStrings.declareGrid(
         ` columnWidth="200" `,
-        '', ColumnDefinitions.idNameJobHireDate)
+        '', ColumnDefinitions.idNameJobHireDate, '', '<igx-paginator *ngIf="paging"></igx-paginator>')
 })
 export class ScrollableGridSearchComponent extends BasicGridSearchComponent {
     public data = SampleTestData.generateFromData(SampleTestData.personJobDataFull(), 30);
+    public paging = false;
 }
 
 @Component({
     template: GridTemplateStrings.declareGrid(
         ` columnWidth="200" [height]="null" `,
-        '', ColumnDefinitions.idNameJobTitleCompany)
+        '', ColumnDefinitions.idNameJobTitleCompany, '', '<igx-paginator *ngIf="paging"></igx-paginator>')
 })
 export class GroupableGridSearchComponent extends ScrollableGridSearchComponent {
     public data = SampleTestData.personIDNameJobCompany();
+    public paging = false;
 }
 
 @Component({
@@ -870,13 +873,14 @@ export class GridWithAvatarComponent extends GridWithSizeComponent {
 
 @Component({
     template: `${GridTemplateStrings.declareGrid(`height="1000px"  width="900px" [primaryKey]="'ID'"`, '',
-        ColumnDefinitions.summariesGroupByColumns)}`
+        ColumnDefinitions.summariesGroupByColumns, '', '<igx-paginator *ngIf="paging"></igx-paginator>')}`
 })
 export class SummariesGroupByComponent extends BasicGridComponent {
     public data = SampleTestData.employeeGroupByData();
     public calculationMode = 'rootAndChildLevels';
     public ageSummary = AgeSummary;
     public ageSummaryTest = AgeSummaryTest;
+    public paging = false;
 }
 
 @Component({
@@ -1001,10 +1005,11 @@ export class GridWithPrimaryKeyComponent extends BasicGridSearchComponent {
 
 @Component({
     template: `${GridTemplateStrings.declareGrid(`height="300px"  width="600px" [primaryKey]="'ID'"`, '',
-        ColumnDefinitions.selectionWithScrollsColumns)}`,
+        ColumnDefinitions.selectionWithScrollsColumns, '', '<igx-paginator *ngIf="paging"></igx-paginator>')}`,
 })
 export class SelectionWithScrollsComponent extends BasicGridComponent {
     public data = SampleTestData.employeeGroupByData();
+    public paging = false;
 }
 
 @Component({
@@ -1529,6 +1534,7 @@ export class IgxGridAdvancedFilteringColumnGroupComponent extends BasicGridCompo
         <igx-column width="100px" [field]="'Downloads'" [editable]="true" dataType="number" [header]="'Downloads'"></igx-column>
         <igx-column width="100px" [field]="'Released'" [editable]="true" dataType="boolean" [header]="'Released'"></igx-column>
         <igx-column width="100px" [field]="'ReleaseDate'" [header]="'ReleaseDate'" dataType="date"></igx-column>
+        <igx-paginator *ngIf='grid.paging'></igx-paginator>
     </igx-grid>`
 })
 export class IgxGridClipboardComponent extends BasicGridComponent {
@@ -1558,12 +1564,13 @@ export class DynamicColumnsComponent extends GridWithSizeComponent {
 
 @Component({
     template: `
-    <igx-grid #gridCustomSelectors [primaryKey]="'ID'" [data]="data" [paging]="true" [rowSelection]="'multiple'" [autoGenerate]="false">
+    <igx-grid #gridCustomSelectors [primaryKey]="'ID'" [data]="data" [rowSelection]="'multiple'" [autoGenerate]="false">
         <igx-column width="100px" [field]="'ID'" [header]="'ID'"></igx-column>
         <igx-column width="100px" [field]="'CompanyName'"></igx-column>
         <igx-column width="100px" [field]="'ContactName'" dataType="string"></igx-column>
         <igx-column width="100px" [field]="'ContactTitle'" dataType="string"></igx-column>
         <igx-column width="100px" [field]="'Address'" dataType="string"></igx-column>
+        <igx-paginator></igx-paginator>
         <ng-template igxRowSelector let-rowContext>
             <span class="rowNumber">{{ rowContext.index }}</span>
             <igx-checkbox [checked]="rowContext.selected" (click)="onRowCheckboxClick($event, rowContext)"></igx-checkbox>
@@ -1612,10 +1619,12 @@ export class GridCustomSelectorsComponent extends BasicGridComponent implements 
         <igx-column field="ProductName" header="Product Name" [dataType]="'string'" [editable]="true" [sortable]="true" width="200px">
         </igx-column>
         <igx-column field="OrderDate" header="Order Date" [dataType]="'date'" [editable]="true" width="200px"></igx-column>
+        <igx-paginator *ngIf="paging"></igx-paginator>
     </igx-grid>`
 })
 export class IgxGridRowEditingComponent extends BasicGridComponent {
     public showToolbar = false;
+    public paging = false;
     public data = SampleTestData.foodProductData();
 }
 
@@ -1743,13 +1752,13 @@ export class IgxGridEmptyRowEditTemplateComponent extends BasicGridComponent {
 
 @Component({
     template: `
-    <igx-grid #grid [data]="data" [primaryKey]="'ProductID'" width="900px" height="900px" [rowEditable]="true"
-    [paging]="paging" [perPage]="7">
+    <igx-grid #grid [data]="data" [primaryKey]="'ProductID'" width="900px" height="900px" [rowEditable]="true" >
         <igx-column field="ProductID" header="Product ID" width="150px"></igx-column>
         <igx-column field="ProductName" header="Product Name" [dataType]="'string'" width="200px"></igx-column>
         <igx-column field="InStock" header="In Stock" [dataType]="'boolean'" width="100px"></igx-column>
         <igx-column field="UnitsInStock" header="Units in Stock" [dataType]="'number'" width="150px"></igx-column>
         <igx-column field="OrderDate" header="Order Date" [dataType]="'date'" width="200px"></igx-column>
+        <igx-paginator *ngIf='paging' [perPage]="7"></igx-paginator>
     </igx-grid>`,
     providers: [{ provide: IgxGridTransaction, useClass: IgxTransactionService }],
 })
@@ -1951,7 +1960,7 @@ export class CellEditingScrollTestComponent extends BasicGridComponent {
 @Component({
     template: GridTemplateStrings.declareGrid(
         ` [width]="width" [height]="height" [paging]="'true'" [perPage]="perPage" [primaryKey]="'ProductID'"`,
-        '', ColumnDefinitions.productBasic)
+        '', ColumnDefinitions.productBasic, '', '<igx-paginator></igx-paginator>' )
 })
 export class GridWithUndefinedDataComponent implements OnInit {
     @ViewChild(IgxGridComponent, { static: true })
@@ -2331,6 +2340,7 @@ export class MRLTestComponent {
     <igx-column *ngFor="let c of columns" [sortable]="true" [field]="c.field" [header]="c.field"
         [width]="c.width" [movable]='true' [resizable]='true'>
     </igx-column>
+    <igx-paginator *ngIf="paging"></igx-paginator>
 
     <igx-action-strip #actionStrip>
         <igx-grid-editing-actions [addRow]='true'></igx-grid-editing-actions>
@@ -2350,6 +2360,7 @@ export class IgxAddRowComponent implements OnInit {
 
     public data: any[];
     public columns: any[];
+    public paging = false;
 
     public ngOnInit() {
 
