@@ -19,6 +19,7 @@ import { HIERARCHICAL_SAMPLE_DATA } from 'src/app/shared/sample-data';
                 <igx-column field="ChildLevels" [groupable]="true" [sortable]="true" [editable]="true" [movable]='true'></igx-column>
                 <igx-column field="ProductName" [groupable]="true" [hasSummary]='true' [movable]='true'></igx-column>
         </igx-column-group>
+        <igx-paginator *ngIf="paging"></igx-paginator>
         <igx-row-island [key]="'childData'" #rowIsland [allowFiltering]="true" [rowEditable]="true" [primaryKey]="'ID'">
             <igx-grid-toolbar [grid]="grid" *igxGridToolbar="let grid"></igx-grid-toolbar>
             <igx-column field="ID" [groupable]="true" [hasSummary]='true' [movable]='true'>
@@ -56,6 +57,7 @@ export class IgxHierarchicalGridTestBaseComponent {
 
     public data;
     public pinningConfig: IPinningConfig = { columns: ColumnPinningPosition.Start, rows: RowPinningPosition.Top };
+    public paging = false;
 
     constructor() {
         // 3 level hierarchy
@@ -169,14 +171,16 @@ export class IgxHierarchicalGridRowSelectionNoTransactionsComponent {
 
 @Component({
     template: `
-    <igx-hierarchical-grid [paging]="true" [data]="data" [cellSelection]="false" [height]="'600px'"
+    <igx-hierarchical-grid [data]="data" [cellSelection]="false" [height]="'600px'"
         [width]="'700px'" #hGridCustomSelectors [primaryKey]="'ID'"
         [rowSelection]="'multiple'">
         <igx-column field="ChildLevels"></igx-column>
         <igx-column field="ProductName"></igx-column>
-        <igx-row-island [key]="'childData'" #rowIsland1 [paging]="true" [primaryKey]="'ID'" [rowSelection]="'single'">
+        <igx-paginator></igx-paginator>
+        <igx-row-island [key]="'childData'" #rowIsland1 [primaryKey]="'ID'" [rowSelection]="'single'">
             <igx-column field="ChildLevels"></igx-column>
             <igx-column field="ProductName"></igx-column>
+            <igx-paginator *igxPaginator></igx-paginator>
             <ng-template igxHeadSelector let-headContext>
                 <igx-checkbox [readonly]="true" (click)="handleHeadSelectorClick(headContext)"
                     [checked]="headContext.selectedCount === headContext.totalCount"
@@ -374,14 +378,14 @@ export class IgxHierarchicalGridExportComponent {
         </igx-column-group>
 
         <igx-row-island [key]="'ChildCompanies'" [autoGenerate]="false">
-            <!-- <igx-column-group [movable]="true" [pinned]="false" header="General Information"> -->
+            <igx-column-group [movable]="true" [pinned]="false" header="General Information">
                 <igx-column field="CompanyName" [movable]="true" [sortable]="true" [resizable]="true"></igx-column>
                 <igx-column-group [movable]="true" header="Personal Details">
                     <igx-column field="ContactName" [movable]="true" [sortable]="true" [resizable]="true"></igx-column>
                     <igx-column field="ContactTitle" [movable]="true" [sortable]="true" [resizable]="true"></igx-column>
                 </igx-column-group>
-            <!-- </igx-column-group> -->
-            <!-- <igx-column-group header="Address Information"> -->
+            </igx-column-group>
+            <igx-column-group header="Address Information">
                 <igx-column-group header="Location">
                     <igx-column field="Address" [movable]="true" [sortable]="true" [resizable]="true"></igx-column>
                     <igx-column field="City" [movable]="true" [sortable]="true" [resizable]="true"></igx-column>
@@ -392,7 +396,7 @@ export class IgxHierarchicalGridExportComponent {
                     <igx-column field="Phone" [sortable]="true" [resizable]="true"></igx-column>
                     <igx-column field="Fax" [sortable]="true" [resizable]="true"></igx-column>
                 </igx-column-group>
-            <!-- </igx-column-group> -->
+            </igx-column-group>
         </igx-row-island>
     </igx-hierarchical-grid>
     `
