@@ -120,6 +120,7 @@ import { GridRowAPISampleComponent } from './grid-row-api/grid-row-api.sample';
 import { GridUpdatesComponent } from './grid-updates-test/grid-updates.component';
 import { TreeSampleComponent } from './tree/tree.sample';
 import { GridColumnTypesSampleComponent } from './grid-column-types/grid-column-types.sample';
+import { AccordionSampleComponent } from './accordion/accordion.sample';
 import { GridLocalizationSampleComponent } from './grid-localization/grid-localization.sample';
 import { TreeGridGroupBySampleComponent } from './tree-grid-groupby/tree-grid-groupby.sample';
 import { PaginationSampleComponent } from './pagination/pagination.component';
@@ -129,6 +130,10 @@ const appRoutes = [
         path: '',
         pathMatch: 'full',
         redirectTo: '/avatar'
+    },
+    {
+        path: 'accordion',
+        component: AccordionSampleComponent
     },
     {
         path: 'action-strip',
@@ -502,7 +507,7 @@ const appRoutes = [
     {
         path: 'gridFinJS',
         component: MainComponent
-    },{
+    }, {
         path: 'tree',
         component: TreeSampleComponent
     },
